[IMP] mail: pimp my notification

+++ addons/mail/models/mail_message.py
@@ -76,12 +76,17 @@ class Message(models.Model):
     # recipients
     partner_ids = fields.Many2many('res.partner', string='Recipients')
     needaction_partner_ids = fields.Many2many(
-        'res.partner', 'mail_message_res_partner_needaction_rel', string='Partners with Need Action')
+        'res.partner', 'mail_message_res_partner_needaction_rel', string='Partners with Need Action',
+        domain="[('is_read', '=', False)]")
     needaction = fields.Boolean(
         'Need Action', compute='_get_needaction', search='_search_needaction',
         help='Need Action')
     channel_ids = fields.Many2many(
         'mail.channel', 'mail_message_mail_channel_rel', string='Channels')
+    # notifications
+    notification_ids = fields.One2many(
+        'mail.notification', 'mail_message_id', 'Notifications',
+        auto_join=True, copy=False)
     # user interface
     starred_partner_ids = fields.Many2many(
         'res.partner', 'mail_message_res_partner_starred_rel', string='Favorited By')
@@ -105,9 +110,12 @@ class Message(models.Model):
     @api.multi
     def _get_needaction(self):
         """ Need action on a mail.message = notified on my channel """
-        my_messages = self.sudo().filtered(lambda msg: self.env.user.partner_id in msg.needaction_partner_ids)
+        mymessages = self.env['mail.notification'].sudo().search([
+            ('mail_message_id', 'in', self.ids),
+            ('res_partner_id', '=', self.env.user.partner_id.id),
+            ('is_read', '=', False)]).mapped('mail_message_id')
         for message in self:
-            message.needaction = message in my_messages
+            message.needaction = message in mymessages
 
     @api.multi
     def _is_accessible(self):
@@ -117,8 +125,8 @@ class Message(models.Model):
     @api.model
     def _search_needaction(self, operator, operand):
         if operator == '=' and operand:
-            return [('needaction_partner_ids', 'in', self.env.user.partner_id.id)]
-        return [('needaction_partner_ids', 'not in', self.env.user.partner_id.id)]
+            return ['&', ('notification_ids.res_partner_id', '=', self.env.user.partner_id.id), ('notification_ids.is_read', '=', False)]
+        return ['&', ('notification_ids.res_partner_id', '=', self.env.user.partner_id.id), ('notification_ids.is_read', '=', True)]
 
     @api.depends('starred_partner_ids')
     def _get_starred(self):
@@ -147,9 +155,10 @@ class Message(models.Model):
         """ Remove all needactions of the current partner. If channel_ids is
             given, restrict to messages written in one of those channels. """
         partner_id = self.env.user.partner_id.id
-        if domain is None:
+        delete_mode = not self.env.user.share  # delete employee notifs, keep customer ones
+        if domain is None and delete_mode:
             query = "DELETE FROM mail_message_res_partner_needaction_rel WHERE res_partner_id IN %s"
-            args = [(partner_id,)]
+            args = [(self.env.user.partner_id.id,)]
             if channel_ids:
                 query += """
                     AND mail_message_id in
@@ -162,15 +171,21 @@ class Message(models.Model):
             self.invalidate_cache()
 
             ids = [m['id'] for m in self._cr.dictfetchall()]
+        # elif domain is None and not delete_mode:
+        #     pass
         else:
-            # not really efficient method: it does one db request for the
-            # search, and one for each message in the result set to remove the
-            # current user from the relation.
             msg_domain = [('needaction_partner_ids', 'in', partner_id)]
             if channel_ids:
                 msg_domain += [('channel_ids', 'in', channel_ids)]
             unread_messages = self.search(expression.AND([msg_domain, domain]))
-            unread_messages.sudo().write({'needaction_partner_ids': [(3, partner_id)]})
+            notifications = self.env['mail.notification'].sudo().search([
+                ('mail_message_id', 'in', unread_messages.ids),
+                ('res_partner_id', '=', self.env.user.partner_id.id),
+                ('is_read', '=', False)])
+            if delete_mode:
+                notifications.unlink()
+            else:
+                notifications.write({'is_read': True})
             ids = unread_messages.mapped('id')
 
         notification = {'type': 'mark_as_read', 'message_ids': ids, 'channel_ids': channel_ids}
@@ -192,18 +207,26 @@ class Message(models.Model):
     @api.multi
     def set_message_done(self):
         """ Remove the needaction from messages for the current partner. """
-        partner_id = self.env.user.partner_id
-        messages = self.filtered(lambda msg: partner_id in msg.needaction_partner_ids)
-        if not len(messages):
+        delete_mode = not self.env.user.share  # delete employee notifs, keep customer ones
+
+        notifications = self.env['mail.notification'].sudo().search([
+            ('mail_message_id', 'in', self.ids),
+            ('res_partner_id', '=', self.env.user.partner_id.id),
+            ('is_read', '=', False)])
+
+        if not len(notifications):
             return
-        messages.sudo().write({'needaction_partner_ids': [(3, partner_id.id)]})
+        if delete_mode:
+            notifications.unlink()
+        else:
+            notifications.write({'is_read': True})
 
         # notifies changes in messages through the bus.  To minimize the number of
         # notifications, we need to group the messages depending on their channel_ids
         groups = []
-        current_channel_ids = messages[0].channel_ids
+        current_channel_ids = notifications.mapped('mail_message_id')[0].channel_ids
         current_group = []
-        for record in messages:
+        for record in notifications.mapped('mail_message_id'):
             if record.channel_ids == current_channel_ids:
                 current_group.append(record.id)
             else:
@@ -217,7 +240,7 @@ class Message(models.Model):
 
         for (msg_ids, channel_ids) in groups:
             notification = {'type': 'mark_as_read', 'message_ids': msg_ids, 'channel_ids': [c.id for c in channel_ids]}
-            self.env['bus.bus'].sendone((self._cr.dbname, 'res.partner', partner_id.id), notification)
+            self.env['bus.bus'].sendone((self._cr.dbname, 'res.partner', self.env.user.partner_id.id), notification)
 
     @api.model
     def unstar_all(self):
@@ -270,6 +293,8 @@ class Message(models.Model):
                 partners |= message.partner_ids
             elif not message.subtype_id and message.partner_ids:  # take specified people of message without a subtype (log)
                 partners |= message.partner_ids
+            if message.needaction_partner_ids:  # notified
+                partners |= message.needaction_partner_ids
             if message.attachment_ids:
                 attachments |= message.attachment_ids
             if message.tracking_value_ids:
@@ -311,6 +336,12 @@ class Message(models.Model):
             else:
                 partner_ids = [partner_tree[partner.id] for partner in message.partner_ids
                                 if partner.id in partner_tree]
+            customer_data = []
+            for partner in message.needaction_partner_ids.filtered(lambda partner: partner.partner_share):
+                # TMP
+                notif = message.notification_ids.filtered(lambda notif: notif.res_partner_id == partner)
+                print partner_tree[partner.id][0], partner_tree[partner.id][1], notif
+                customer_data.append((partner_tree[partner.id][0], partner_tree[partner.id][1], notif.is_email, notif.email_state))
             attachment_ids = []
             for attachment in message.attachment_ids:
                 if attachment.id in attachments_tree:
@@ -323,9 +354,11 @@ class Message(models.Model):
             message_dict.update({
                 'author_id': author,
                 'partner_ids': partner_ids,
+                'customer_data': customer_data,
                 'attachment_ids': attachment_ids,
                 'tracking_value_ids': tracking_value_ids,
             })
+            print message_dict['customer_data']
 
         return True
